port to 1.5.1

+++ Umamusume/Model/Account.cs
@@ -21,7 +21,7 @@ namespace Umamusume.Model
         public Account() : this(Guid.NewGuid()) { }
 
         public Guid Udid { get; set; }
-        public int ViewerId { get; set; }
+        public long ViewerId { get; set; }
         [JsonIgnore]
         public string SessionId => Utils.Bin2Hex(Utils.MakeMd5($"{ViewerId}{Udid}"));
         public string Authkey { get; set; } = null;
