+++ Umamusume/Model/Requests/CircleItemRequest.cs
@@ -11,7 +11,7 @@ namespace Umamusume.Model
 
 
 
-        public int? viewer_id;
+        public long? viewer_id;
