+++ Umamusume/Model/Requests/TeamStadiumUserDetailRequest.cs
@@ -5,7 +5,7 @@ namespace Umamusume.Model
     {
 
 
-        public int? target_viewer_id;
+        public long? target_viewer_id;
